requirements.txt
requests
python-dotenv
langchain-groq
langchain-core
langchain-community
sentence-transformers
faiss-cpu
streamlit
speechrecognition
pydub
langdetect
deep-translator
gtts
groq
PyPDF2
langchain-text-splitters
flask
flask-cors